move abstractions back to a separate library
