+++ WCountLogic/WCountLib.Abstractions/Counters/IByteCounter.cs
@@ -11,14 +11,13 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WCountLib.Counters.Abstractions
+namespace WCountLib.Abstractions.Counters
 {
-    public interface ICharCounter
+    public interface IByteCounter
     {
-        public int CountCharacters(string s, Encoding textEncodingType);
+        public int CountBytes(string s, Encoding textEncodingType);
         
-        public ulong CountCharacters(IEnumerable<string> enumerable, Encoding textEncodingType);
-        public Task<ulong> CountCharactersAsync(IEnumerable<string> enumerable, Encoding textEncodingType);
-
+        public ulong CountBytes(IEnumerable<string> strings, Encoding textEncodingType);
+        public Task<ulong> CountBytesAsync(IEnumerable<string> strings, Encoding textEncodingType);
     }
 }
