+++ WCountLogic/WCountLib.Abstractions/Counters/ILineCounter.cs
@@ -7,16 +7,14 @@
     file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-// ReSharper disable InconsistentNaming
-
 using System.Collections.Generic;
 
-namespace WCountLib.Detectors.Abstractions
+namespace WCountLib.Abstractions.Counters
 {
-    public interface IWordDetector
+    public interface ILineCounter
     {
-        bool IsStringAWord(string s, bool excludeStringsWithSpaces = true);
-        bool IsStringAWord(string s, IEnumerable<char> delimitersToExclude, bool excludeStringsWithSpaces = true);
+        int CountLines(string s);
 
+        int CountLines(IEnumerable<string> enumerable);
     }
 }
