+++ WCountLogic/WCountLib/Counters/CharCounter.cs
@@ -13,7 +13,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WCountLib.Counters.Abstractions;
 using WCountLib.Localizations;
 
 // ReSharper disable RedundantIfElseBlock
